First working demo of camera stream/display
The demo performs low-latency video streaming
from a camera to an HTTP server via mjpg_streamer.

A python script reads from the local server
and displays the image.

There are also setup instructions in the README.md

This was tested on my desktop with the ELP 2.8mm USB camera
module that I bought from amazon:
http://www.amazon.com/gp/product/B00KA80N9U?psc=1&redirect=true&ref_=oh_aui_detailpage_o02_s00

The camera is rated for mjpg 1280x720 @ 60fps
The display client read that it was receiving frames at
about 50fps.

Next steps is to test this on the Odroid to see if this
frame rate will still hold up.